move context into module
